app-utils: Move msghdr_x and recvmsg_x declaration from client
